pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "firebase-admin>=7.1.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "google-genai>=1.36.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
